+++ src/print_pointer.c
@@ -1,6 +1,6 @@
 #include "../include/ft_printf.h"
 
-int	print_pointer(void *pointer, int *flags)
+int	print_pointer(void *pointer, size_t *flags)
 {
 	return (print_hexadecimal((t_ull) pointer, flags, 'x'));
 }
